Changed $HOME to $GAMEDIR
This script was encouraging users to change the $HOME variable if the game was not located in $HOME/proto. The tweaked $HOME variable would be inherited by the game executable at runtime and could unexpectedly creep into shell commands spawned by the game server. (i.e. autoarchive)

To fix the issue, I've adjusted the script to use a safer $GAMEDIR variable that defaults to $HOME. If the user customizes the value of this variable, we won't end up with unexpected behavior later.